Put back the baseline for cspell, so only new words are triggered
We see a lot of remarks being thrown in our PR's now, which is insightfull but way to noisy and blocking.
By putting back the baseline at least it doesn't block our PR's.
We can easily migrate correct words from our baseline to the dictionary and after remove the baseline again.

But this way it's to disruptive.