Update search to take profile field visibility into account
This introduces a new ProfileFieldVisibilityProcessor which uses the
global visibility settings and the per-user visibility settings for
users to index content in the Search API with the visibility in mind.
This allows us to limit searches only to those visibility types that a
searching user has access to.

As a bonus we can now remove the RestrictedNameProcessor which did
something similar but only for the first name/last name and nickname.
This is now properly handled by the broader per-field permission system.