Ensure profile creation works with registration form
To make our code simpler we require that a profile exists and use
hook_user_insert to make sure this happens regardless of how the user is
created. However there was an issue in the profile module that caused a
profile to be saved after use creation and its related hooks were
completed.

Depending on how we responded to that it would cause the profile not to
be created at all or it would cause a duplicate profile to be created
which then still failed being default.

This change includes a patch to optimize the way profiles are created in
the profile module during user registration which also ensures the
profile is available during welcome e-mail sending.