Ensure unverified users can't list users through search
We want to protect the overview of users on platforms which is why we
introduced the `list user` permission which is also coming from Drupal
issue #2953566. Unverified users were already not allowed to view the
user search but they were still allowed to search for users through the
"all" search which is inconsistent.

This commit ensures that unverified users also can't use the all search
to enumerate users. We need a separate check for this with respect to
anonymous users because unverified users are able to see user profiles
with the fields that can be seen by unverified users whereas anonymous
users can't view profiles at all.