Remove "Limit Real Name" functionality
This was originally built because customers wanted real names to be
public while still providing the user with the ability to hide their
real name behind a nickname. However implementing this same logic in the
current system has become quite complicated.

Users now also have the ability to control the visibility of their
profile fields more specifically which allows them to also hide their
real name behind a nickname if the site managers give them the ability
to do so.

We ensure an upgrade path is included to make sure the behaviour doesn't
unexpectedly change.

This change reveals that the autocomplete still needs work to ensure it
works with the new permission system. We should probably change the
autocomplete to use the user search instead.