Clean up references to old group types
This doesn't remove config objects in its entirety but instead fixes
configurations that are broken by the removal of the old groups. We can
see that some shared views that are specific for our group types can be
moved to the flexible groups module because they're specific to how
groups work in Open Social and not generic to anything using `group`.