Replace `access users search` permission with `list user`
In a9f41dbd69253ac09c26d3f74b6d8d31be931359 we ensured that search
queries required the `list user` permission so that search can not be
used to enumerate users. This was needed because otherwise users might
show up on the "All" search page.

The user search still had a custom permission. However we don't actually
expose a configuration page for that permission. So we change the view
to use the permission that is actually needed in the underlying data
system and we remove the special purpose permission.