Remove previous profile tag split code
We had initially deprecated the SocialProfileTagSplitWidget in 12.0.0
for removal in 13.0.0. However, it relies on the SocialProfileTagService
for which we have already migrated and cleaned up the config, so in
practice those systems already don't work.

I think it's better to remove the code and have a hard-break than keep
the code around in a non-functioning matter creating subtle bugs for
people. I also don't want anyone to accidentally build on or extend the
old system.

Unfortunately it is not possible to deprecate these things in 11.x
already because the replacement can't be introduced until 12.0.0. This
makes the break a bit harder but we'll write an update guide to help out
developers.