Ensure profile always exists regardless of hook order
This has been quite the challenge. There is code that's unaware of the
existence of profiles, but that inadvertently triggers code that
requires profiles. For example the `ginvite` module needs to send
notifications when a user is created and does this in
`hook_user_insert`. However this includes the display name of the
account to address the user. We want to do this with their name if
they've filled out this information on the registration page.

We use module_implements_alter to change the order in which hooks fire
to make sure the profile and social_profile hooks are always run first.
This ensures whenever a user is saved, a profile will exist (together
with the profile patch added earlier) and makes sure the user's name is
properly generated if available.

A backtrace is added to logging because figuring all this out without it
was a pain.