Display tag items as pills
We already had this styling in our FollowTag formatting, but it's
actually a universal design pattern for tags in Open Social that they're
displayed as small pills. To accomplish this we add the classes in our
TagSplitFormatter so that themes can use this.

This is accompanied by a field file change in socialbase that ensures
the main label for the profile tag field is not shown but the individual
labels of the subfields are rendered as labels instead.