Make "Verified user" label uniform
When the role was introduced we installed new sites with "Verified user"
capitalization but the update introduced it with "Verified User". In
case the roles are used in labels that are clicked in tests, these
differences in capitalization are problematic, and it also just doesn't
look the way we intend for a large part of our customerbase.

Our designer gave the following answer: "Based on our microcopy approach
(Slite/Humans/Product language) we don’t use Title Case, which means
that the right option is ‘Verified user’."

This commit updates existing sites and pulls them in line with the
correct spelling already used in new sites.