Fix profile tagging display in socialbase
It's a bit weird to add a patch for something we manage. However we have
a chicken/egg problem. We can't add the new twig file to socialbase
because it'll break the existing profile tagging functionality. However,
we require it to properly display the new profile tagging display
structure and pass our tests.

We include the patch until the profile refactor is released, then we can
tag a new socialbase version with the fix and update within Open Social
while removing the patch.