Revert search_api_language filter ID change in search views
In #3470 we changed the search_api_language filter to a
language_with_fallback filter to ensure that users could find matching
content in different languages if no version in their own language was
available. However, this also changed the IDs of the filter which can
break existing sites that might rely on that filter.

In this commit we revert the ID/name change and only change the field
used. We also revert any formatting changes made to the views.* files to
make fixing merge conflicts in our search rework easier.