Backport patch from Drupal Core from issue #3231503 with cache improvements for hook_entity_extra_field_info() and EntityFieldManager::getExtraFields()
As this patch only lands in 10.1.x (which is when we can remove this patch) we backport it as it could have significant performance impact, especially
on pages layout builder or paragraphs driven.
This way we can get the benefits already from the work done there, and remove it once we do our update towards 10.1.x, which is not yet planned at the time or writing.