Split content tagging search tests
We split the search tests up per type of entity we're searching against.
This keeps the feature files a bit shorter and easier to read.

We also found a few copy paste errors in the user tests that are fixed.

Finally we tweak the tests and add a new scenario for each type to
search for untagged things which is also supported in the new widget
type. The tests are modified to encode the expected behaviour for the
widget in the test as well (since it's configuration dependent).