Change empty text and field ID for content tagging facet
We want to add more facets to search so we change the ID and the
untagged label to distinguish more easily from other facet fields within
the same search.

No update path is needed because this is a fix for an unreleased piece
of code. No update path is changed because the config/modify and
optional config are only enabled when the facets module is enabled which
is in the upgrade path that's not yet released/executed.