Fix view name for groups search
We can't just fix the change in `social_search_update_11403` because we
would leave sites that had already executed the update in an
indeterminate state.

There's two scenarios:

1) The update has already run. In this case the change for the groups
view hasn't happened due to the type, the new update hook will be
executed and now all views are correct.

2) The update hook hasn't run already. In this case we do the correct 2
views (that sites from scenario 1 don't need help changing) in 11403 and
we fix groups in 11404.