Fix expected config in `social_event_update_11902`
We're expecting a specific key to exist, but testing showed that it's
possible the keys in the list are -1, 0 and 1, and not 0, 1, and 2. In
that case the expected config fails. To resolve this we remove the
expectation of a specific commit and make sure we fix the config
regardless of what's currently stored in it.