Reorder social_tagging upgrade paths
We were already using the `social_tag_split` formatter, but we moved the
code to a new module afterwards. We need to be sure that the module is
enabled before we use the plugins so we don't run into errors.

Reordering these update hooks is safe because the code has not been
released yet.