Ensure TagSplitWidget doesn't render when empty
The widget was rendering an empty container when it had no actual
subfields to show. This also causes form group wrappers to show
up which made the profile page look weird with only an empty group.

This commit ensures the field wrapper isn't rendered in case there are
no sub fields to render for selection which correctly hides the field
group wrapper too.