Issue #3384622: Check if permissions of type "translate $bundle $entity_type_id" are valid.
As of Drupal 10, we can no longer have invalid permissions
in the database (https://www.drupal.org/node/3193348).
Permissions of the type "translate $bundle $entity_type_id"
are valid if the given bundle is configured as translatable.
If bundle is not translatable, permission is removed
from all existing roles. On clean installs, permissions
of type "translate $bundle $entity_type_id" will need to be
manually enabled as of this commit.

Note that this commit was cherry picked from
3a682174595b35b934f12e9a104185d63084f0cd and
we have to change the code to resolve conflict with
social_group_flexible_group_update_111101
defined twice, as this was falsely introduced on
4024a9760f7aeec859cfa6ad2cd153bb7b9ac33a.