Dump deserialized mail to spool directory for failed tests
It's already helpful that the serialized mail is made available to
developers in the mail-spool, which allows it to be output from the CI
for easier debugging. However this requires developers to go through a
serialized PHP object which can be difficult.

This commit ads a postScenario hook that runs when a test fails and
splits out the e-mail into multiple parts: html, text, and headers.
These are stored in separate files which makes them a lot easier to
inspect for developers, for example opening the HTML file in a browser.
In case they need information that isn't split out, the original
serialized object is still available in the .message file.