Ensure all users can't see profile edit page if they can't edit any fields
We had previously written code that prevented users from needing to have
"edit any profile" permission if they could potentially edit a field on
other people's profile (e.g. content managers) to ensure they didn't see
empty profiles.

This is now being extended to "edit own profile" to cover the scenario
where all fields are disabled or users are not allowed to edit any of
their own profile fields. We discussed with the design team that it's
desirable to not show the profile edit call to actions in case there are
no editable fields. The profile field settings are assumed to be mostly
static in the lifetime of a platform, so it's unlikely that users will
have profile edit access and then at some point lose it. Instead we
expect that in cases where profile data is externally managed, they will
not have the profile edit access to begin with. The alternative of
showing a message on the empty profile edit page was considered to be a
worse experience because it leads the user to a dead-end and might
encourage questions to site managers even though the site managers have
purposefully configured the platform in a certain way.

A new "All profile fields are <status>" step is added to the
ProfileContext since listing all the fields ran into situations where
optional modules introduced fields that the list could not rely on
(because it would fail the "without optional" test run), but that
were enabled by default when the optional module was enabled. As a bonus
it makes the test a lot more readable.

Because we must take the field status into account to fix an issue where
a person might have edit permission but the field isn't on the platform
anymore, we must also ensure that roles are saved if the field status is
changed. This leads to a change in SocialProfileSettingsForm and a
similar temporary workaround in our ProfileContext for testing.

The `add any profile profile` permission is removed everywhere because
while working on this we found out that that permission doesn't actually
exist. The correct permission is `create profile profile`, but this is
not a permission we need: the system will create the default profile for
everyone and we don't want more than 1 profile of our 'profile' type.

We also remove the `update any profile profile` and 'update own profile
profile' assignment for sitemanagers and authenticated users
respectively, as this will now be taken care of by our role_presave hook
implementation.

Finally we add test coverage for user creation through the
administration screen since we discovered that there was a bug in the
interaction with the profile widget and all screens being disabled. This
required a patch to core to fix behaviour in the ManagedFile element
which was causing a required field error even for fields a user could
not access.

We also change the profile edit test for all fields being disabled
because in our CI test run with optional modules it's possible that
there are field which are not enabled by default. Disabling all fields
in a list would then cause those fields to still be enabled and the test
to fail. A new test step to disable all fields helps us work around
this.