Issue #3384622: Check if permissions of type "translate $entity_type_id" are valid.
As of Drupal 10, we can no longer have invalid permissions
in the database (https://www.drupal.org/node/3193348).
Permissions of the type "translate $entity_type_id" are
valid if the given bundle is configured as translatable.
If bundle is not translatable, permission is removed
from all existing roles. On clean installs, permissions
of type "translate $entity_type_id" will need to be
manually enabled as of this commit.