Restructure EmailContext to provide better debug information
The EmailContext mail handling was created quite some years ago. In case
an email in a test didn't match, the developer only got information that
there was no match, there was no useful information about what didn't
match (for example that a subject wasn't found, or that a specific line
didn't match).

The test steps are rewritten so that the developer gets as much
actionable information as possible. This can be that a specific subject
wasn't found, or that one or more lines of the body did not match. This
makes it easier for the developer based on just the test output and the
changes they may have made to determine where the error comes from.

Providing this information in the test output requires some new
functions to be rewritten. This opportunity is used to also restructure
the other functions of the file and provide PHPStan output:

- Behat hooks are placed at the top
- Followed by developer usable test steps
- Followed by protected functions which are used to support the test
  steps, in the order in which they were first called.

Dump deserialized mail to spool directory for failed tests

It's already helpful that the serialized mail is made available to
developers in the mail-spool, which allows it to be output from the CI
for easier debugging. However this requires developers to go through a
serialized PHP object which can be difficult.

This commit ads a postScenario hook that runs when a test fails and
splits out the e-mail into multiple parts: html, text, and headers.
These are stored in separate files which makes them a lot easier to
inspect for developers, for example opening the HTML file in a browser.
In case they need information that isn't split out, the original
serialized object is still available in the .message file.