Optimize menu link loading for module installs
A pattern that was seen to add the "All x" links was that we'd load all
the menu links (which can be 100s) and then looped through them all.

This commit instead uses filters in the database query to more
efficiently find the correct link and then more efficently creates the
new child menu link. While making this change we move things back into
hook_install, which is generally small enough not to need splitting up
(especially since we usually don't re-use any of the other install code,
since installing is a pretty specific action).