Issue #3386004: Allow installing dynamic_entity_reference ^3
Dynamic Entity Reference 1.16 does not provide compatibility with Drupal
10 and unfortunately version 3.0 immediately drops support for Drupal 9,
which would make upgrading to it a breaking change.

We allow both versions so that Composer can select the correct version
depending on what version of Drupal is installed.

Our live is slightly complicated because we're affected by the bug
https://drupal.org/node/3099176 which requires incompatible patches
between 1.16 and 3.0, removing the patch is unfortunately not an
option.

We introduce the goalgorilla/open_social_upgrade_patches package which
we will release in a 1.x version and a 2.x version. The 1.x version will
only contain patches that are _only_ compatible with modules for Drupal
9 and the 2.x version will contain patches that are _only_ compatible
with Drupal 10. That providees us with a small in-between that composer
can change between depending on the installed Drupal or module version
to select the correct patches that might be required.

We depend on the dev version of goalgorilla/open_social_upgrade_patches
so that we can work on it until we tag a new stable Open Social release
at which point we should give it a version number so that tagged Open
Social versions depend on specific package versions.

Note that version `^1.16 || ^3` is used and not `^1 || ^2` for two
reasons: we need a minimum of 1.16 for dynamic_entity_reference to make
sure the 1.x patch applies and other bugs/features are fixed/available.
Additionally the 2.x module was released with a change to the entity
storage and it has a matching 4.x version with D10 support. Reference
the module's project page for more info on how its versions relate.