After updating VBO to the latest version a while ago, the types were never updated.
The request now returns Error: Argument of an invalid type bool|float|int|string|null supplied for foreach, only iterables are supported.
Which we cant iterate over, looking at the code it was heavily copied from VBOController.
While we can even further improve it useing the new trait, I don't want to do that for regression right now
so we just re-use the logic provided by VBOController to ensure it is in sync but we still show our customized ajax response message.