Lock our behat tests to run from 11.9.0 for our update path testing.
Right now we have 12.0.x opened. However it's not tagged yet with a stable tag.
Which means we are testing from 10.x up to 12.0.x, unfortunately that results in errors.
This is because in order to update to Drupal 10, you need to be at least on a version with Drupal 9.4, which is what we've shipped in 11.x
That's why we're now locking the version to 11.9.0, we can unlock it again later when we're working on 13.x for example.