We need to add a patch to add Drupal 10 compatibility in the color info yml file.
It was already part of the composer.json file, but not in the info.yml file.
After checking https://www.drupal.org/project/project_composer/issues/3084063#comment-13414884
it shows the composer.json file should take precedense so this does not need to use the lenient plugin as it's more Drupal related rather than composer related.