Add test module for group migration
We want to be able to test the group migration with a sufficient number
of groups but it's cumbersome to create these all the time. This commit
provides a module so that we can test the database migration for the
groups but creating groups and content in an update hook before the
migration and then validating the results in an update hook after the
migration.

This test module showed two bugs which we immediately fix:

1. Because update hooks all run together and we make the database
queries directly, none of the cache invalidation logic for entities
runs. This means we need to invalidate the Redis/Database (depending on
back-end config) cache and the static in-memory cache, manually.

2. Testing of the `field_group_description` showed that the field data
for groups was not yet properly migrated. Steps have been added so that
the field tables also have update queries. We combine this with a
validation that all groups on the old group types exist on the flexible
group entity to help developers who might have customised the old group
types do the same on the flexible group type.