Update open_social_dev to 2.1
The 2.0 version of the package should've been needed for Drupal 10
compatibility but apparently in the distribution that works fine with 1.4.0.

Nevertheless we should upgrade to 2.0.0 to make sure we update Drush
and all of our tooling.

We immediately jump to 2.1.0 because it adds `kingdutch/cucumber-linter`
which we want to use for linting our Behat tests.

1.4.0 is still allowed so that we can install the previous major version
of Open Social which still requires PHP 7.4 which is not supported by
Drush 12.