Fix PHPStan errors due to newer PHPStan-Drupal version
We fix the errors we receive or that have changed text but we keep the
errors that affect our group related modules since we know that work is
ongoing in those files in other PRs that should fix these issues.

We also ignore an "always true" which can become dangerous if removed
due to an overeager PHPStan-Drupal check.

See https://drupal.slack.com/archives/C033S2JUMLJ/p1701250878367139