Issue #3406164 - Add our own inline-css plugin to make sure our email template styling gets converted correctly.
We are having an issue where the inline-css plugin does not see any css and will not go through the css-inline plugin and will not add inline css. This causes gmail to not have any styling.
As a workaround we are adding our own plugin who basically does the same apart from that it will do this always, not only if there is additional css.
Additionally we put a fallback in place, which adds `mail_html` as format option, as there are scenario's where there is no format, which means the emails will be escaped.
Mail_html feels like a safe fallback there.