Improve Behat support for SOLR
When running tests backed by SOLR locally you may run into content
still being in the SOLR database when the database gets reset. This
is because our Behat test doesn’t clear out the SOLR database when a
new MariaDB database is loaded, but we can easily do this with
onDatabaseLoaded.

It can also be challenging when writing search Behat tests to make
sure that text you’re searching for is part of the result set and
doesn’t trigger on the search input or on a filter. For this we want
to add two steps that actually look inside of the content region to
exclude those other sections.

Finally we currently have a “are the search indexes up to date” step
which has a custom implementation. However we can get the code for
free by invoking Drush instead. That reduces the amount of code we
need to maintain and ensures that if indexing changes in the Search
API module we get those changes for free (assuming the Drush API
doesn’t change).