Bump Behat CI previous major version installation to PHP8.1
For Open Social 10 and earlier and a large part of Open Social 11 we
needed to use PHP 7.4 for the installation (since code wasn't ready for
newer PHP versions). However in the latest Open Social 11 release
there's accidentally a bit of code that requires PHP 8.1.

This was fixed in 97b88234e76e42c327ba8808f8ecb71a5da403bf but that
wasn't backported to an 11.x release.

We know that going forward all the previous major versions that we want
to test from will support at least PHP 8.1 so we can bump the PHP
version we use to install the previous Open Social major version to
solve the problem. That is since we will not create any new 11.x
releases anymore.