Test Behat from latest stable major version to `main`
We've made changes to our testing set-up in which `main` is the next
major version rather than the current major version. So rather than
testing from the last major version to the current major version, we
accept that that path works and test instead from the current stable
major version to the one in development.

This does leave us with a need to find a different solution for ensuring
backports don't break other upgrade paths.