Ensure Behat tests abort if updates fail with requirements errors
Drush requires the -y flag to run in automated environments because it
prompts with the list of available updates. However it also does an
update requirements check which gets skipped by the same -y option.

This means that our update path in the CI doesn't fail if there are
requirements errors (even though it should, because it causes errors
later in tests that are hard to debug) and our tests against fixture
dumps for specific update scenarios also don't fail.

Until Drush provides a better solution in
https://github.com/drush-ops/drush/issues/5806. We manually go through
the update output and errors if the key sentence for requirements errors
is spotted.