Remove upgrade path test for user following
We moved the user setting "allow to follow me" from the Profile entity
to the User entity. The upgrade path is now being removed, assuming that
all sites have gone through it or are newer installs that no longer
needed. This corresponds with cleaning up of all the older update hooks.