PROD-28057: Add optional string and phone number GraphQL input types
These types can be used in system to system API design to distinguish
between a field that a system doesn't want to change or a field that
should be changed to the empty value (NULL).

For now we only need String and PhoneNumber, we don't have this need for
other scalar or more complex types yet.