Add additional composer install
We somehow run in to a scenario, where patches added to our Distro composer.json
are not applied directly after our composer require. This also happens locally.
A second run of composer install seems to do the trick.

After looking at composer-patches we've noticed in the new 2.0.0 (still in beta atm)
this will be solved using a patches lock file and better management of patches.

For now, considering time criticality we are adding the second composer install run.
We've attempted to clear caches before the composer commands in the CI,
adding composer-patches to the test/composer.json and more.

Composer install running twice uses caches still so is much faster as well.