In: https://www.drupal.org/project/drupal/issues/2253257 a modal for entity delete operation was added.
Unfortunately this doesn't look great across the product, and was sneaked in to the core release.
So we revert the changes from Core for the time being, this gives us the opportunity to go through our design phase
and make sure we can make this a consistent improvement across the product.