Better deal with our data policy creation
We used to test a weird edge case, where the machine name seemingly wasn't filled in. That would likely throw a validation error.
But that shouldn't be part of our test in the shape it was. Also that isn't appearing anymore after updating to 10.2 of Drupal Core.
So instead, we now write the machine name and remove the validation work around. This should keep it the same value wise.