Error suppression for file_mdm exif metadata issues.
Although this needs to have a fix upstream (see related issue) it seems that's far from available. The attached patch is the one from the merge request from the linked issue made by the maintainer, which isn't ideal just removing it, but considering it's the maintainers choice for now perhaps our best bet especially as it's linked to the issue in the contrib module, which makes it easier for us to track it's progress.

Manual testing (and the fact it returns NULL on error) seems like there isn't much of a downside to suppress the error for now, alternatively if this stays open for long without an upstream fix we could decide to change the log level to warning perhaps. But the fact of the matter is that these images are coming from user input and we have no way of helping them fix the metadata of the image, nor can they do it themselves, it's just an issue with the modules interpretation of it.