Remove our custom views bulk operations callback, causing several issues
After updating to views_bulk_operations 4.2.6, we noticed the actual selection wasn't updating anymore correctly
this blocked the bulk operations action button. On top of that we have a selection message which talks about members
but this callback is used for every bulk operations, including content and more. We need beter ways of fixing it.
So instead we deprecate this callback, and move to ViewsBulkOperationsController's variant.