Issue #3421348: Remove redundant group type filters from group views
For views.view.group_events we remove specific references to group
types. To ensure we only show event content we make the relationship to
the events required. We can also remove the validation of the group
type.

The view is also moved to `config/optional` so that it’s only available
if the views module and the social_event module are enabled. This brings
us closer to not requiring a dependency on `social_event` from
`social_group`. The other changes are relatively small but out of scope
for this issue.

For views.view.group_topics we make similar changes, but we add
`social_topic` as a dependency for the view since it contains the
required node bundle. Again this is moved to `config/optional` in a
first step to decouple `social_group` from `social_topic`.

For views.view.group_manage_members we also remove the filters. There
might be multiple different relationships between a user and a group.
For example “invites”, “membership requests” and actual “memberships”.
The `gc__user` relationship ensures that we only show actual memberships.
This also provides an option to unify the invites/request/member views
by playing with these relationships.

The same goes for views.view.group_members. However, here the
relationship doesn’t exist yet so we add it to ensure we only show
full-members.

For views.view.group_managers we make use of the naming for group roles
being `<group_type>-<role_name>` and the fact that we use `group_manager`
as role name consistently across our existing group types. This allows
us to match for a role `-group_manager` regardless of which group the
role belongs to. While we’re at it we simplify the groupings of this
filter.

There are other views that we leave as non-optional for future changes.

SocialGroupAdminPeopleConfigOverride is removed because the modifications
it makes to views are no longer needed.

The relevant parts of the Flexible Group and Secret Group Config
Overrides are also removed because they cause errors in the filters.