Refactor activity stream test to use flexible group
The activity-stream test is a pretty old test that was still using the
UI to create all the scaffold content needed. It also used the group add
menu for creating a group but that breaks with the removal of the old
group types so we create a flexible group directly instead.