Internal: Ensure there are no install errors
Drush doesn't actually exit if there's an error during installation
which is really annoying. We found out because our CI was passing even
though the installer threw errors.

To fix this we apply a similar fix which we previously applied in the
update steps which captures and checks the output text to check if
there's any errors.