Move flexible group specific field from search to flexible group module
In the migration from old group types to flexible groups we added a
field to the group search. However that field doesn't exist if flexible
groups isn't enabled and the search will be installed before flexible
groups. So to make sure the search config is valid we need to make sure
that the flexible group module adds the field rather than it being in
the social_search module.

We don't yet move the entire index and view for two reasons:

1) There's plans to merge the group index into `all` at which point
we'll also have to move the field indexing, but that's out of current
scope and already in motion elsewhere (profile refactor).

2) We want to have the `group` view in `flexible_group` because the view
should only show flexible groups. However, currently it's also still
used to show group-like things such as `courses`. Before limiting the
group view to Flexible Groups we must make sure courses has its own way
of being found in search by creating a separate view. That's too much
work for the amount of time we have at the moment.