Run GitHub Behat Workflow tests against SOLR 9
We still need 8.11 to install in the previous version of Open Social.
However when updating to `main` or the branch under test we need to
start using SOLR 9 which we also use in the update and for new installs.

This commit updates the image for the Behat runner jobs which was
forgotten. It leaves 8.11 in the "install previous version" jobs.