Revert solr in github actions/tugboat back to 8.11
We tried ugprading our CI to SOLR 9 but for some reason Drupal times out
the connection during install against SOLR 9. It works locally but we
can not figure out why it doesn't work within GitHub.

There were some changes in how the SOLR image is created between SOLR 8
and 9 since the project moved to the Apache repo. So that'll take some
deeper investigation than we currently have time for.