Disable group activity stream test
The functionality for showing comments on activity stream items for
flexible groups isn't implemented yet so we temporarily disable that
test and provide a test to at least prove that activities for content
creation appear on group activity streams.

We add a small replacement test to cover at least the part of the
functionality that currently works and prevent regressions. A follow-up
is created to fix the larger functionality.