After adding page cache, behat is sometimes returning cached pages
This is usually for paths we visit (or have visited before)
Considering we're having our own implementation of the Chromedriver visit
we do a force reload in here after visiting a page, just to make sure
we don't get a cached page. We're hoping we will find a better solution in the future.