Restart the session in a before scenario
This means the browser cache is cleared and not shared between scenario's.
This is necessary due to enabling cache.page.max_age which adds the Cache Control header and allows the browser.
This leads to different scenario's where the cache is shared for AN users, in the browser, between scenario's.
For example, visiting the group overview or home page with different config changes didn't take any effect that way.