Internal: Improve test fixture creation for group references
This commit allows group reference data to be provided as a comma
separated list of IDs or term names (mixing is allowed). This can now be
used in tables for e.g. groups, topics and events. By putting this in
the validate method we can use this everywhere.

This ensures we deduplicate the group title lookup so that we don't
break on a pre-fetched value. Since we now do it by field type we no
longer need to do it by field name.