Issue #3403634: Fix access denied for anonymous users on group canonical path
The group path `/group/{group}` redirects users to the the group's Stream page (`/group/{group}/stream`). If the group is Public, all users — including `anonymous` — should have access to this route.

Since this is a redirect route, the access should be checked in the destination route. This commit changes the route's access requirements to `_access: 'TRUE'`, making it accessible for all.